requirements.txt
anyio==3.6.2
certifi==2023.5.7
click==8.1.3
dnspython==2.3.0
email-validator==2.0.0.post2
fastapi==0.95.2
h11==0.14.0
httpcore==0.17.1
httptools==0.5.0
httpx==0.24.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
orjson==3.8.12
pydantic==1.10.7
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.5.0
ujson==5.7.0
uvicorn==0.22.0
uvloop==0.17.0  #not available for windows
watchfiles==0.19.0
websockets==11.0.3
sqlacodegen==2.3.0.post1
psycopg2-binary
sqlalchemy==1.4.48
python-jose==3.3.0
passlib==1.7.4
redis==4.5.5